+++ src/palace/manager/service/integration_registry/catalog_services.py
@@ -6,12 +6,14 @@ from palace.manager.integration.goals import Goals
 from palace.manager.service.integration_registry.base import IntegrationRegistry
 
 if TYPE_CHECKING:
-    from palace.manager.marc.exporter import MarcExporter  # noqa: autoflake
+    from palace.manager.integration.catalog.marc.exporter import (  # noqa: autoflake
+        MarcExporter,
+    )
 
 
 class CatalogServicesRegistry(IntegrationRegistry["MarcExporter"]):
     def __init__(self) -> None:
-        from palace.manager.marc.exporter import MarcExporter
+        from palace.manager.integration.catalog.marc.exporter import MarcExporter
 
         super().__init__(Goals.CATALOG_GOAL)
         self.register(MarcExporter)
