+++ tests/manager/api/controller/test_marc.py
@@ -7,8 +7,8 @@ import pytest
 from flask import Response
 
 from palace.manager.api.controller.marc import MARCRecordController
+from palace.manager.integration.catalog.marc.exporter import MarcExporter
 from palace.manager.integration.goals import Goals
-from palace.manager.marc.exporter import MarcExporter
 from palace.manager.service.integration_registry.catalog_services import (
     CatalogServicesRegistry,
 )
